requirements.txt
asgiref==3.6.0
Django==4.2
django-taggit==3.1.0
Faker==18.6.0
Pillow==9.5.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
